Create own Event class. TODO : repair EventManager.

+++ core/android/src/main/java/com/toxic/prikupa/android/AndroidContext.java
@@ -25,7 +25,7 @@ public class AndroidContext implements Context {
   }
 
   @Override
-  public String convertMillseconds(float millsec) {
+  public String convertMillseconds(double millsec) {
     return format.format(new Date((long) millsec));
   }
